modify

+++ GameProject/Project/Engine/MeshData.h
@@ -17,7 +17,7 @@ public:
     virtual void Load(const wstring& _strFilePath);
     virtual void Save(const wstring& _strFilePath);
     Ptr<CMaterial> GetMtrl(int _iIdx) { return m_vecMtrl[_iIdx]; }
-    UINT Size() { return m_vecMtrl.size(); }
+    UINT Size() { return (UINT)m_vecMtrl.size(); }
 
     Ptr<CMesh> GetMesh() { return m_pMesh; }
     vector<Ptr<CMaterial>> GetMaterial() { return m_vecMtrl; }
